Change `ShowBloodSugarValidationError` effect as view effect
