+++ app/src/main/java/org/simple/clinic/bloodsugar/entry/BloodSugarEntryUiActions.kt
@@ -1,5 +1,6 @@
 package org.simple.clinic.bloodsugar.entry
 
+import org.simple.clinic.bloodsugar.BloodSugarMeasurementType
 import org.simple.clinic.bloodsugar.BloodSugarUnitPreference
 import java.util.UUID
 
@@ -12,4 +13,13 @@ interface BloodSugarEntryUiActions {
   fun setBloodSugarSavedResultAndFinish()
   fun showConfirmRemoveBloodSugarDialog(bloodSugarMeasurementUuid: UUID)
   fun showBloodSugarUnitSelectionDialog(bloodSugarUnitPreference: BloodSugarUnitPreference)
+  fun showBloodSugarEmptyError()
+  fun showBloodSugarHighError(
+      measurementType: BloodSugarMeasurementType,
+      unitPreference: BloodSugarUnitPreference
+  )
+  fun showBloodSugarLowError(
+      measurementType: BloodSugarMeasurementType,
+      unitPreference: BloodSugarUnitPreference
+  )
 }
