+++ app/src/main/java/org/simple/clinic/bloodsugar/entry/BloodSugarEntryViewEffectHandler.kt
@@ -1,5 +1,6 @@
 package org.simple.clinic.bloodsugar.entry
 
+import org.simple.clinic.bloodsugar.BloodSugarUnitPreference
 import org.simple.clinic.mobius.ViewEffectsHandler
 
 class BloodSugarEntryViewEffectHandler(
@@ -16,6 +17,21 @@ class BloodSugarEntryViewEffectHandler(
       is SetBloodSugarSavedResultAndFinish -> uiActions.setBloodSugarSavedResultAndFinish()
       is ShowConfirmRemoveBloodSugarDialog -> uiActions.showConfirmRemoveBloodSugarDialog(viewEffect.bloodSugarMeasurementUuid)
       is ShowBloodSugarUnitSelectionDialog -> uiActions.showBloodSugarUnitSelectionDialog(viewEffect.bloodSugarUnitPreference)
+      is ShowBloodSugarValidationError -> showBloodSugarValidationError(viewEffect.result, viewEffect.unitPreference)
+    }
+  }
+
+  private fun showBloodSugarValidationError(
+      result: ValidationResult,
+      unitPreference: BloodSugarUnitPreference
+  ) {
+    when (result) {
+      ValidationResult.ErrorBloodSugarEmpty -> uiActions.showBloodSugarEmptyError()
+      is ValidationResult.ErrorBloodSugarTooHigh -> uiActions.showBloodSugarHighError(result.measurementType, unitPreference)
+      is ValidationResult.ErrorBloodSugarTooLow -> uiActions.showBloodSugarLowError(result.measurementType, unitPreference)
+      is ValidationResult.Valid -> {
+        /* no-op */
+      }
     }
   }
 }
